% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodata.R
\name{CMPI6_bioclim_metadata}
\alias{CMPI6_bioclim_metadata}
\title{Retrieve the CMIP6 variable metadata}
\usage{
CMPI6_bioclim_metadata()
}
\value{
a tibble of metadata
}
\description{
Retrieve the CMIP6 variable metadata
}
\seealso{
\href{https://www.worldclim.org/data/bioclim.html}{BioClim}
}
