% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocache-package.R
\docType{package}
\name{geocache}
\alias{geocache}
\title{geocache: Simple wrapper around the geodata package}
\description{
Wraps the geodata package in a manner to permit local storage of data and to make xyzt package friendly.
}
