% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{geocache_path}
\alias{geocache_path}
\title{Retrieve the geocache data path}
\usage{
geocache_path(..., root = geocache_root())
}
\arguments{
\item{...}{char, path segments}

\item{root}{char, the root directory specification}
}
\value{
path specification, untestes for existence
}
\description{
Retrieve the geocache data path
}
