% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodata.R
\name{SSP_metadata}
\alias{SSP_metadata}
\title{Retrieve a table of SSP metadata}
\usage{
SSP_metadata()
}
\value{
tibble of SSP metadata
}
\description{
Retrieve a table of SSP metadata
}
\seealso{
\href{https://en.wikipedia.org/wiki/Shared_Socioeconomic_Pathways}{Shared Socioeconomic Pathways}
}
